chore: major refactor
- Split index.tsx into separate components for better organization
- Rename HelloWorld to SetupSync to better reflect its purpose
- Move App component to App/index.tsx
- Move SetupSync component to App/SetupSync.tsx
- Wrap async operations in IIFEs for proper error handling
- Clean up imports and remove unused ones
- Maintain existing functionality while improving code structure
