+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "MyReactNativeApp",
+  "name": "CNPlusSync",
   "version": "0.0.1",
   "private": true,
   "scripts": {
